+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/SearchHighlighter.java
@@ -132,7 +132,7 @@ public final class SearchHighlighter {
   void nextSearchPosition() {
     if (!searchTextPositions.isEmpty()) {
       deselectCurrentSearchPosition();
-      if (currentSearchTextPositionIndex.isNull() || currentSearchTextPositionIndex.equalTo(searchTextPositions.size() - 1)) {
+      if (currentSearchTextPositionIndex.isNull() || currentSearchTextPositionIndex.isEqualTo(searchTextPositions.size() - 1)) {
         currentSearchTextPositionIndex.set(0);
       }
       else {
@@ -148,7 +148,7 @@ public final class SearchHighlighter {
   void previousSearchPosition() {
     if (!searchTextPositions.isEmpty()) {
       deselectCurrentSearchPosition();
-      if (currentSearchTextPositionIndex.isNull() || currentSearchTextPositionIndex.equalTo(0)) {
+      if (currentSearchTextPositionIndex.isNull() || currentSearchTextPositionIndex.isEqualTo(0)) {
         currentSearchTextPositionIndex.set(searchTextPositions.size() - 1);
       }
       else {