+++ swing/framework-model-tools/src/main/java/is/codion/swing/framework/model/tools/explorer/DefinitionRow.java
@@ -23,10 +23,12 @@ import is.codion.framework.domain.entity.EntityDefinition;
 public final class DefinitionRow {
 
   final DatabaseDomain domain;
+  final String tableType;
   final EntityDefinition definition;
 
-  DefinitionRow(DatabaseDomain domain, EntityDefinition definition) {
+  DefinitionRow(DatabaseDomain domain, String tableType, EntityDefinition definition) {
     this.domain = domain;
+    this.tableType = tableType;
     this.definition = definition;
   }
 }