+++ swing/framework-model-tools/src/main/java/is/codion/swing/framework/model/tools/metadata/MetadataColumn.java
@@ -52,7 +52,7 @@ public final class MetadataColumn {
     this.decimalDigits = decimalDigits;
     this.nullable = nullable;
     this.defaultValue = defaultValue;
-    this.comment = comment;
+    this.comment = comment == null ? null : comment.trim().replace("\"", "\\\"");
     this.primaryKeyIndex = primaryKeyIndex;
     this.foreignKeyColumn = foreignKeyColumn;
     this.autoIncrement = autoIncrement;