+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityPanel.java
@@ -1058,30 +1058,30 @@ public class EntityPanel extends JPanel {
   //#############################################################################################
 
   private void requestEditPanelFocus() {
-    if (editPanelState.equalTo(HIDDEN)) {
+    if (editPanelState.isEqualTo(HIDDEN)) {
       editPanelState.set(EMBEDDED);
     }
     editPanel().requestInitialFocus();
   }
 
   private void selectInputComponent() {
-    if (editPanelState.equalTo(HIDDEN)) {
+    if (editPanelState.isEqualTo(HIDDEN)) {
       editPanelState.set(EMBEDDED);
     }
     editPanel().selectInputComponent();
   }
 
   private void updateEditPanelState() {
-    if (editPanelState.notEqualTo(WINDOW)) {
+    if (editPanelState.isNotEqualTo(WINDOW)) {
       Window editPanelWindow = parentWindow(editControlPanel);
       if (editPanelWindow != null) {
         editPanelWindow.dispose();
       }
     }
-    if (editPanelState.equalTo(EMBEDDED)) {
+    if (editPanelState.isEqualTo(EMBEDDED)) {
       editControlTablePanel.add(editControlPanel, BorderLayout.NORTH);
     }
-    else if (editPanelState.equalTo(HIDDEN)) {
+    else if (editPanelState.isEqualTo(HIDDEN)) {
       editControlTablePanel.remove(editControlPanel);
     }
     else {
@@ -1093,10 +1093,10 @@ public class EntityPanel extends JPanel {
   }
 
   private void toggleEditPanelState() {
-    if (editPanelState.equalTo(WINDOW)) {
+    if (editPanelState.isEqualTo(WINDOW)) {
       editPanelState.set(HIDDEN);
     }
-    else if (editPanelState.equalTo(EMBEDDED)) {
+    else if (editPanelState.isEqualTo(EMBEDDED)) {
       editPanelState.set(WINDOW);
     }
     else {
@@ -1137,7 +1137,7 @@ public class EntityPanel extends JPanel {
 
     @Override
     public void execute() {
-      if (containsEditPanel() && editPanelState.equalTo(HIDDEN)) {
+      if (containsEditPanel() && editPanelState.isEqualTo(HIDDEN)) {
         editPanelState.set(WINDOW);
       }
     }